refactor(studio.giselles.ai): Move @/components/v2/ui to app/(main)/settings
These components are used in settings section only.
