+++ apps/studio.giselles.ai/app/(main)/settings/team/agent-usage-dialog.tsx
@@ -1,3 +1,4 @@
+import { Button } from "../components/button";
 import {
 	Dialog,
 	DialogContent,
@@ -5,7 +6,6 @@ import {
 	DialogTitle,
 	DialogTrigger,
 } from "@/components/ui/dialog";
-import { Button } from "@/components/v2/ui/button";
 import { type AgentActivity, AgentUsageTable } from "./agent-usage-table";
 
 type AgentUsageDialogProps = {
