+++ apps/studio.giselles.ai/app/(main)/settings/team/billing-section.tsx
@@ -1,6 +1,6 @@
+import { Button } from "../components/button";
 import { Card } from "@/components/ui/card";
 import { Skeleton } from "@/components/ui/skeleton";
-import { Button } from "@/components/v2/ui/button";
 import { fetchCurrentTeam, isProPlan } from "@/services/teams";
 import { manageBilling } from "@/services/teams/actions/manage-billing";
 import { upgradeTeam } from "@/services/teams/actions/upgrade-team";
